nitpicker.py: nest rule loop under line loop
This fixes line ordering in the output.

Say line 1 matches rule B and line 2 matches rule A.  Before this commit
we would see the print for line 2 first and the print for line 1 later.
This is so because rule B comes later in the rule list than rule A.

With this patch the line numbers will always be in sequence irrespective
of the order of the rules.

Signed-off-by: Abdur Rehman <arehmanq199@gmail.com>